Split CLI parsing tests into per-command-area modules
